Rename ObjectTraverseFunctions into ObjectTraverser
