+++ src/Mock.ts
@@ -6,7 +6,7 @@ import {MethodStub} from "./stub/MethodStub";
 import {ReturnValueMethodStub} from "./stub/ReturnValueMethodStub";
 import {strictEqual} from "./ts-mockito";
 import {MockableFunctionsFinder} from "./utils/MockableFunctionsFinder";
-import {ObjectTraverser} from "./utils/ObjectTraverseFunctions";
+import {ObjectTraverser} from "./utils/ObjectTraverser";
 import {ObjectPropertyCodeRetriever} from "./utils/PrototypeKeyCodeGetter";
 
 export class Mocker {
